Break up integration tests into success and fail files
